requirements.txt
GitPython
bandit
click
requests
pyyaml
